Adds GET Function and Adds Component
This adds a function to the index.jsx file that can call the GET
request to the server. This will need to be done in the lifecycle so
that it can be added on refresh.

It also adds a new SingleRepo component that will load all repos as
they are received back from the server.